blob_upload_cloudrun/requirements.txt
flask==3.0.2
google-cloud-storage==2.16.0
